Improved readability
I tried a single letter styling for this project the first time I made it. I don't like this, it's really hard to understand intuitevely. I change it to use a personal version of BEM. In the future, when it's more refined, I might provide documentation for the BEM system I use.